[WIP] storage/auth_tokens: refactor methods
